Still working project logic
