+++ frontend/assets/js/projectCards.js
@@ -7,6 +7,9 @@
 
 
 import { AREA_MAPPING, STATUS_MAPPING } from './constants.js';
+import api from './api.js';
+import { fetchProjects,fetchUsers} from './api.js';
+import { populateFormForEdit } from './projectForm.js';
 
 
 /**
@@ -17,7 +20,7 @@ import { AREA_MAPPING, STATUS_MAPPING } from './constants.js';
     * @returns {Array} The formatted array of project objects.
     */
 //TODO: Change username to name when model is updated
-export default function formatProjects(projects, users) {
+export function formatProjects(projects, users) {
     return projects.map(project => {
         const techLead = users.find(user => user.id === project.tech_lead);
         const businessLead = users.find(user => user.id === project.business_lead);
@@ -29,6 +32,87 @@ export default function formatProjects(projects, users) {
             status: STATUS_MAPPING[project.status] || 'Unknown'
         };
     });
+}
+
+
+
+export function renderProjects(projects) {
+    console.log("Rendering projects", projects);
+    const projectsList = document.getElementById('projects-list');
+    projectsList.innerHTML = '';
+    projects.forEach(project => {
+        projectsList.appendChild(createProjectCard(project));
+    });
+}
+
+
+/**
+ * Creates a project card element for a given project.
+ * 
+ * @param {Object} project - The project object to create a card for.
+ * @returns {HTMLElement} The created project card element.
+ */
+export function createProjectCard(project) {
+    const card = document.createElement('div');
+    card.className = 'project-card';
+    
+    const statusClass = `status-${project.status.toLowerCase().replace(" ","-")}`;
+    const dates = formatProjectDates(project.start_date, project.end_date);
+
+    card.innerHTML = `
+        <div class="project-card-header">
+            <h3 class="project-card-title">${project.project_name}</h3>
+            <span class="project-status ${statusClass}">${capitalizeFirstLetter(project.status)}</span>
+        </div>
+        <div class="project-card-body">
+            ${project.description || 'No description provided'}
+            <div class="project-details">
+                <div class="detail-item">
+                    <strong>Area:</strong> ${capitalizeFirstLetter(project.area)}
+                </div>
+                <div class="detail-item">
+                    <strong>Tech Lead:</strong> ${project.techLead}
+                </div>
+                <div class="detail-item">
+                    <strong>Business Lead:</strong> ${project.businessLead}
+                </div>
+            </div>
+        </div>
+        <div class="project-card-footer">
+            <span>${dates}</span>
+            <div class="project-actions">
+                <button class="edit-project" data-id="${project.id}">Edit</button>
+                <button class="delete-project" data-id="${project.id}">Delete</button>
+            </div>
+        </div>
+    `;
+
+    card.querySelector('.delete-project').addEventListener('click', async () => {
+        if (confirm(`Are you sure you want to delete this project?`)) {
+
+            //Remove the project from the DB and refresh the page/projects
+            const response = await api.delete(`projects/delete/${project.id}/`);
+            if (response.status === 204) {
+                console.log('Project deleted successfully');
+               let projects = await fetchProjects();
+               let users = await fetchUsers();
+               projects = formatProjects(projects, users);
+               renderProjects(projects);
+            }
+
+
+        }
+    });
+
+    card.querySelector('.edit-project').addEventListener('click', () => {
+        let editingProjectId = project.id;
+        populateFormForEdit(project);
+        document.querySelector('.submit-button').textContent = 'Update Project';
+        document.querySelector('.project-form-container h2').textContent = 'Edit Project';
+        projectForm.scrollIntoView({ behavior: 'smooth' });
+    });
+
+    return card;
 }
 
 
@@ -73,65 +157,4 @@ function formatProjectDates(start, end) {
     if (!start) return `Due by ${formatDate(end)}`;
     if (!end) return `Started ${formatDate(start)}`;
     return `${formatDate(start)} - ${formatDate(end)}`;
-}
-
-/**
- * Creates a project card element for a given project.
- * 
- * @param {Object} project - The project object to create a card for.
- * @returns {HTMLElement} The created project card element.
- */
-export function createProjectCard(project) {
-    const card = document.createElement('div');
-    card.className = 'project-card';
-    
-    const statusClass = `status-${project.status.toLowerCase().replace(" ","-")}`;
-    const dates = formatProjectDates(project.start_date, project.end_date);
-
-    card.innerHTML = `
-        <div class="project-card-header">
-            <h3 class="project-card-title">${project.project_name}</h3>
-            <span class="project-status ${statusClass}">${capitalizeFirstLetter(project.status)}</span>
-        </div>
-        <div class="project-card-body">
-            ${project.description || 'No description provided'}
-            <div class="project-details">
-                <div class="detail-item">
-                    <strong>Area:</strong> ${capitalizeFirstLetter(project.area)}
-                </div>
-                <div class="detail-item">
-                    <strong>Tech Lead:</strong> ${project.techLead}
-                </div>
-                <div class="detail-item">
-                    <strong>Business Lead:</strong> ${project.businessLead}
-                </div>
-            </div>
-        </div>
-        <div class="project-card-footer">
-            <span>${dates}</span>
-            <div class="project-actions">
-                <button class="edit-project" data-id="${project.id}">Edit</button>
-                <button class="delete-project" data-id="${project.id}">Delete</button>
-            </div>
-        </div>
-    `;
-
-    card.querySelector('.delete-project').addEventListener('click', () => {
-        if (confirm(`Are you sure you want to delete this project?`)) {
-            projects = projects.filter(p => p.id !== project.id);
-            fetchProjects();
-            saveProjects();
-
-        }
-    });
-
-    card.querySelector('.edit-project').addEventListener('click', () => {
-        editingProjectId = project.id;
-        populateFormForEdit(project);
-        document.querySelector('.submit-button').textContent = 'Update Project';
-        document.querySelector('.project-form-container h2').textContent = 'Edit Project';
-        projectForm.scrollIntoView({ behavior: 'smooth' });
-    });
-
-    return card;
 }
